+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Converted Shipit to use the new GitHub Apps API, older installations will have to redo their setup.
+
 # 0.21.0
 
 * Dropped MRI 2.2 support since it's EOL.